requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
django-crispy-forms==1.6.1
django-filter==1.0.2
djangorestframework==3.6.3
gunicorn==19.7.1
mysqlclient==1.3.10
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
